// packages/client/src/store/settings/actionTypes.ts
export enum SettingsActionTypes {
  SetViewInFiatCurrency = 'SETTINGS.SET_VIEW_IN_FIAT_CURRENCY',
  SetCryptoCurrency = 'SETTINGS.SET_CRIPTYO_CURRENCY',
  ChangePassword = 'SETTINGS.CHANGE_PASSWORD',
  ChangePasswordResponse = 'SETTINGS.CHANGE_PASSWORD_RESPONSE',
  ChangeEmail = 'SETTINGS.CHANGE_EMAIL',
  ChangeEmailResponse = 'SETTINGS.CHANGE_EMAIL_RESPONSE',
  Get2FA = 'SETTINGS.GET_2FA',
  Set2FA = 'SETTINGS.SET_2FA',
  Set2FAError = 'SETTINGS.SET_2FA_ERROR',
  Set2FAActivate = 'SETTINGS.SET_2FA_ACTIVATE',
  set2FADeactivate = 'SETTINGS.SET_2FA_DEACTIVATE',
  Set2FAActivateError = 'SETTINGS.SET_2FA_ACTIVATE_ERROR',
  set2FADeactivateError = 'SETTINGS.SET_2FA_DEACTIVATE_ERROR',
  set2FAActivateResponce = 'SETTINGS.SET_2FA_ACTIVATE_RESPONSE',
  set2FADeactivateResponce = 'SETTINGS.SET_2FA_DEACTIVATE_RESPONSE',
  SetPrivacy = 'SETTINGS.SET_PRIVACY',
}